[C8] Split sketch.go by responsibility (#2245)
model/sketch/sketch.go was 655 LOC, over the 500-line ceiling. Split the 2D
sketch aggregate along its responsibilities; each file is now under the
ceiling:

- sketch.go (286): the Sketch aggregate — the struct, entity/collection
  accessors, the constraint-collection accessors, the plane/frame binding and
  initCollections/ToModel/ToSketch.
- sketch_base.go (160): the Entity interface and the embedded base struct
  every entity carries (id/name/edit/visible/shared flags, health, display
  style) plus the process-wide id sequence.
- sketch_entities.go (120): entity mutation — add/remove/delete and the
  drop-from-collection, plus the sketch-point allocation arena.
- sketches.go (123): the Sketches collection (create named/auto-named sketches
  on a plane, shared parameters/block definitions, lookup/remove).

Behaviour-preserving: split programmatically by receiver-qualified
declaration, verified every non-comment code line is byte-identical (only
per-file import redistribution differs); imports resolved per file with
goimports. Sketch package builds + tests pass, archguard passes, lint 0. Net
delta of recognizers / tolerance constants / fallback sites / type assertions:
0.

Part of #2183 (C8: structure/size limits). Closes #2245.

(cherry picked from commit a5a384b66718fdf37978b8d3aa6db7b3d0acfb75)
